Add nearly comparison support for collections
Use nearly comparison (epsilon, ulps, tolerance and default) on collections
with any type T, as long as T implements the nearly traits.
The tolerance values for the collection of T will be the same as for type T.

This implements the collections: array, slice, Vec, VecDeque and LinkedList
as well as several combinations of these.

This commit also changes the library structure. All nearly traits and
implementations are now in the file nearly_eq.rs instead of having the
traits and implementations for eps, ulps, tolerance and default in
separate files.

All tests for collection combinations are in the file collection_test.rs.
The former array and slice tests have been removed from the other test files.
The other test files now only test the primitive types f32 and f64.